% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.datascan_report}
\alias{print.datascan_report}
\title{Print a datascan report}
\usage{
\method{print}{datascan_report}(x, ..., top_n = 5)
}
\arguments{
\item{x}{A datascan_report}

\item{...}{Unused}

\item{top_n}{Number of issues to show.}
}
\description{
Print a datascan report
}
