% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/details.R
\name{report_details}
\alias{report_details}
\title{Extract details from a datascan report}
\usage{
report_details(report)
}
\arguments{
\item{report}{A \code{datascan_report}.}
}
\value{
The \code{details} list stored in the report.
}
\description{
Extract details from a datascan report
}
