% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/details.R
\name{report_issues}
\alias{report_issues}
\title{Extract issues from a datascan report}
\usage{
report_issues(report)
}
\arguments{
\item{report}{A \code{datascan_report}.}
}
\value{
A data.frame of issues.
}
\description{
Extract issues from a datascan report
}
