% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/details.R
\name{report_meta}
\alias{report_meta}
\title{Extract meta from a datascan report}
\usage{
report_meta(report)
}
\arguments{
\item{report}{A \code{datascan_report}.}
}
\value{
The \code{meta} list stored in the report.
}
\description{
Extract meta from a datascan report
}
