% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{scan}
\alias{scan}
\title{Scan an R object and report potential data issues}
\usage{
scan(x, ...)
}
\arguments{
\item{x}{An R object (e.g. data.frame, vector)}

\item{...}{Reserved for future use}
}
\value{
A datascan_report object
}
\description{
Scan an R object and report potential data issues
}
